Add project seldon to the showcase
We've been building this tool for a while but just starting to open it up and
post about it more publicly. Would love for anyone to try it out and give us
some feedback! TypeSense has been working like charm, even in the places
where we're probably abusing its intended use case 😁