Add featured on Unzip.dev badge
I've written a developer trend issue about Search as a service on Unzip.dev and found about typesense. I created a badge that will show you are one of the key players in the field. I personally, as the creator of Unzip, chose typesense as the personal recommendation to my readers.

Unzip.dev is a developer trends newsletter with over 2200 tech founders (and growing). Hopefully this badge PR will be helpful for potential users of typesense 🙏